Only detect edges where needed
This is an optimisation suggested by @cgearing where the edge detection
is only run on the regions of the image where it is needed.

This is coming about because I want to get the edge detection running
fast enough that it can run fairly frequently on a mobile phone. As it's
being used to detect if a playing card is properly framed we only need
to run the detection on that frame.